[feature] Forever mode. Requests logs endlessly
Add a option to requests logs in an endless loop, repeating the requests
at a constant interval.

The interval is specified through the --interval option. Which defaults
to 10 minutes. time.Tick is used, which adjusts the intervals or drops
ticks to make up for slow requests or writers.

Considering that logshare logs take up to 30 minutes to be available,
the --offset option have been added to specify how much behind to
requests logs from. Defaults to 30 minutes.